move GlobalConfiguration Descriptotr out of Builder
